Rename Commands namespace to Command
- Singular form is consistent with other namespaces
